
\name{deprecateddata}
\alias{deprecateddata}

\title{
Deprecated data sets in CaseCohortCoxSurvival
}
\description{
These data sets still work but will be removed (defuncted) in the next version of the package.

\code{dataexample} is deprecated and will be removed in the next version of the package.

\code{dataexample.missingdata} is deprecated and will be removed in the next version of the package.

}
%\details{ }
%\examples{}
%\keyword{data}
\seealso{\code{\link{dataexample.stratified}}, \code{\link{dataexample.unstratified}}, \code{\link{dataexample.missingdata.stratified}}, \code{\link{dataexample.missingdata.unstratified}}}



